refactor(tangle-cloud): move AccountStatsCard from component to pages
